sources/requirements.txt
# SPDX-FileCopyrightText: 2022 Orthanc Team SRL <info@orthanc.team>
#
# SPDX-License-Identifier: GPL-3.0-or-later

pyjwt==2.4.0
fastapi==0.78.0
jinja2==3.1.2
pytz==2022.1
python-dateutil==2.8.2
uvicorn[standard]==0.18.2
pytest
httpx